Resume interrupted carryover rollbacks ahead of the migration ladder
